requirements.txt
ftfy==5.3.0
python-escpos==2.2.0
python-opengraph==0.0.2
CherryPy==17.0.0
